requirements.txt
#Package==Version
#Python 3.12.2
#--------------------------
azure-core==1.30.0
numpy==1.26.4
openai==1.12.0
openpyxl==3.1.2
pandas==2.2.0
psycopg2==2.9.9
pyarrow==15.0.0
python-dotenv==1.0.1
requests==2.31.0
sqlalchemy==2.0.27

# TOTAL PIP LIB @ DEVELIVERY TO TK IT Department 
# annotated-types    0.6.0
# anyio              4.2.0
# asttokens          2.4.1
# azure-core         1.30.0
# certifi            2024.2.2
# charset-normalizer 3.3.2
# colorama           0.4.6
# comm               0.2.1
# debugpy            1.8.1
# decorator          5.1.1
# distro             1.9.0
# et-xmlfile         1.1.0
# executing          2.0.1
# h11                0.14.0
# httpcore           1.0.3
# httpx              0.26.0
# idna               3.6
# ipykernel          6.29.2
# ipython            8.21.0
# jedi               0.19.1
# jupyter_client     8.6.0
# jupyter_core       5.7.1
# matplotlib-inline  0.1.6
# nest-asyncio       1.6.0
# numpy              1.26.4
# openai             1.12.0
# openpyxl           3.1.2
# packaging          23.2
# pandas             2.2.0
# parso              0.8.3
# pip                24.0
# platformdirs       4.2.0
# prompt-toolkit     3.0.43
# psutil             5.9.8
# psycopg2           2.9.9
# pure-eval          0.2.2
# pyarrow            15.0.0
# pydantic           2.6.1
# pydantic_core      2.16.2
# Pygments           2.17.2
# python-dateutil    2.8.2
# python-dotenv      1.0.1
# pytz               2024.1
# pywin32            306
# pyzmq              25.1.2
# requests           2.31.0
# setuptools         69.1.0
# six                1.16.0
# sniffio            1.3.0
# stack-data         0.6.3
# tornado            6.4
# tqdm               4.66.2
# traitlets          5.14.1
# typing_extensions  4.9.0
# tzdata             2024.1
# urllib3            2.2.0
# wcwidth            0.2.13
# wheel              0.42.0